Layout an igraph object using easylayout's web application.

The easylayout function takes an igraph object and serializes` it into a web application integrated with the IDE’s interface through a Shiny server. The web application lays out the network by simulating attraction and repulsion forces. Simulation parameters can be adjusted in real-time. An editing mode allows moving and rotating nodes. Once the user finishes tinkering the layout, it is sent back to the R session to be plotted through popular libraries like ggplot2 or even the base package itself.

Usage

easylayout(graph)

Arguments

graph

An igraph object representing the network to be laid out.

Value

A two column matrix with XY coordinates and N rows, where N is the number of vertices in the graph.

Examples

if (FALSE) { # \dontrun{
library(igraph)
g <- make_ring(10)
g <- easylayout(g)
plot(g)
} # }